Edited track horizontalvelocity and tests to reflect reality.

+++ ATMSystem/ATMSystem/Interfaces/ITrack.cs
@@ -8,7 +8,7 @@ namespace ATMSystem.Interfaces
         ICoordinate CurrentPosition { get; set; }
         ICoordinate LastKnownPosition { get; set; }
         int CurrentAltitude { get; set; }
-        int CurrentHorizontalVelocity { get; set; }
+        double CurrentHorizontalVelocity { get; set; }
         int CurrentCompassCourse { get; set; }
         void Update(ICoordinate coord, int altitude, DateTime timestamp);
         DateTime LastSeen { get; set; }
